+++ src/long.lua
@@ -289,6 +289,16 @@ end
 Long.__add = Long.add
 
 --[[
+ * Returns the bitwise AND of this Long and the specified.
+ * @param {!Long|number|string} other Other Long
+ * @returns {!Long}
+--]]
+function Long:band(other)
+  if not Long.isLong(other) then other = Long.fromValue(other) end
+  return Long.fromBits(bit32.band(self.low, other.low), bit32.band(self.high, other.high), self.unsigned)
+end
+
+--[[
  * Returns the bitwise NOT of this Long.
  * @returns {!Long}
 --]]